fix: moved `BlockInputs` into the block producer
