+++ proto/src/lib.rs
@@ -9,7 +9,6 @@ pub mod generated;
 
 pub use domain::{
     accounts::{AccountInputRecord, AccountState},
-    blocks::BlockInputs,
     convert, nullifier_value_to_block_num,
     nullifiers::NullifierWitness,
     transactions::TransactionInputs,
